\name{dbmss-package}
\alias{dbmss-package}
\alias{dbmss}
\docType{package}
\title{
  Distance-Based Measures of Spatial Structures
}
\description{
  Simple computation of spatial statistic functions of distance to characterize spatial structures of mapped objects, including classical ones (Ripley's \emph{K} and others) and more recent ones used by spatial economists (Duranton and Overman's \emph{Kd}, Marcon and Puech's \emph{M}). It relies on spatstat for some core calculation.
}
\details{
\tabular{ll}{
Package: \tab dbmss\cr
Type: \tab Package\cr
Version: \tab 2.1.2\cr
Date: \tab 2014-11-01\cr
License: \tab GNU General Public License\cr
}
}
\author{
Eric Marcon,
Gabriel Lang,
Stephane Traissac,
Florence Puech

Maintainer: Eric Marcon <Eric.Marcon@ecofog.gf>
}
\references{
  Marcon, E., Lang, G., Traissac, S. and F. Puech (submitted). Tools to Characterize Point Patterns: dbmss for R. \emph{Journal of Statistical Software}.
}
\keyword{ package }
