\name{is.wmppp}
\alias{is.wmppp}
\title{
  Test whether an object is a weighted, marked, planar point pattern
}
\description{
  Check whether its argument is an object of class {"wmppp"} (\code{\link{wmppp.object}}).
}
\usage{
is.wmppp(X)
}
\arguments{
  \item{X}{
  Any object
  }
}
\value{
  \code{TRUE} if \code{X} is a weighted, marked, planar point pattern, otherwise \code{FALSE}.
}
\author{
  Eric Marcon <Eric.Marcon@ecofog.gf>
}
\seealso{
  \code{\link{wmppp.object}}
}

