\name{dbmss-package}
\alias{dbmss-package}
\alias{dbmss}
\docType{package}
\title{
  Distance based measures of spatial structures
}
\description{
  This package allows simple computation of a full set of spatial statistic functions of distance, including classical ones (Ripley's \emph{K} and others) and more recent onces used by spatial economists (Duranton and Overman's \emph{Kd}, Marcon and Puech's \emph{M}). It relies on spatstat for core calculation.
}
\details{
\tabular{ll}{
Package: \tab dbmss\cr
Type: \tab Package\cr
Version: \tab 1.2.4\cr
Date: \tab 2012-11-27\cr
License: \tab GNU General Public License\cr
}
}
\author{
Eric Marcon,
Gabriel Lang,
Stephane Traissac,
Florence Puech

Maintainer: Eric Marcon <Eric.Marcon@ecofog.gf>
}
\references{
Marcon, E. and F. Puech (2012). A typology of distance-based measures of spatial concentration. \emph{HAL SHS}. 00679993.
}
\keyword{ package }
