\name{dbmss-package}
\alias{dbmss-package}
\alias{dbmss}
\docType{package}
\title{
  Distance based measures of spatial structures
}
\description{
  This package allows simple computation of a full set of spatial statistic functions of distance to characterize spatial structures of mapped objects, including classical ones (Ripley's \emph{K} and others) and more recent ones used by spatial economists (Duranton and Overman's \emph{Kd}, Marcon and Puech's \emph{M}). It relies on spatstat for core calculation.
}
\details{
\tabular{ll}{
Package: \tab dbmss\cr
Type: \tab Package\cr
Version: \tab 2.0.3\cr
Date: \tab 2013-05-24\cr
License: \tab GNU General Public License\cr
}
}
\author{
Eric Marcon,
Gabriel Lang,
Stephane Traissac,
Florence Puech

Maintainer: Eric Marcon <Eric.Marcon@ecofog.gf>
}
\references{
  Marcon, E. and F. Puech (submitted). Tools to Characterize Point Patterns: dbmss for R. \emph{Journal of Statistical Software}.
}
\keyword{ package }
