\name{dbmss-package}
\alias{dbmss-package}
\alias{dbmss}
\docType{package}
\title{
  Distance Based Measures of Spatial Structures
}
\description{
  Simple computation of spatial statistic functions of distance to characterize the spatial structures of mapped objects, including classical ones (Ripley's \emph{K} and others) and more recent ones used by spatial economists (Duranton and Overman's \emph{Kd}, Marcon and Puech's \emph{M}). Relies on spatstat for some core calculation.
}
\author{
Eric Marcon,
Gabriel Lang,
Stephane Traissac,
Florence Puech

Maintainer: Eric Marcon <Eric.Marcon@ecofog.gf>
}
\references{
  Marcon, E., and Puech, F. (2003). Evaluating the Geographic Concentration of Industries Using Distance-Based Methods. \emph{Journal of Economic Geography}, 3(4), 409-428.

  Marcon, E. and Puech, F. (2010). Measures of the Geographic Concentration of Industries: Improving Distance-Based Methods. \emph{Journal of Economic Geography} 10(5): 745-762.
  
  Marcon, E., Puech F. and Traissac, S. (2012). Characterizing the relative spatial structure of point patterns. \emph{International Journal of Ecology} 2012(Article ID 619281): 11.
  
  Lang G., Marcon E. and Puech F. (2014) Distance-Based Measures of Spatial Concentration: Introducing a Relative Density Function. \emph{HAL} 01082178, 1-18.

  Marcon, E., Traissac, S., Puech, F. and Lang, G. (2015). Tools to Characterize Point Patterns: dbmss for R. \emph{Journal of Statistical Software}. 67(3): 1-15.
  
  Marcon, E. and Puech, F. (2017). A typology of distance-based measures of spatial concentration. \emph{Regional Science and Urban Economics}. 62:56-67.
}
\keyword{package}
