\name{KtoL}
\alias{KtoL}
\title{
	Nomalization of \emph{K}-like functions into \emph{L}-like functions
}
\description{
  This function is used internally to transform \emph{K} into \emph{L}.
}
\usage{
KtoL(Value, r)
}
\arguments{
  \item{Value}{
  A vector of values.
 }
  \item{r}{
  The vector of distances.
 }
}
\value{
  A vector of values.
}
\author{
	Eric Marcon <Eric.Marcon@ecofog.gf>
}

\keyword{internal}
